pyproject.toml
[project]
name = "subtitle-contrast-prototype"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "numpy>=2.2",
    "pillow>=12.0",
    "fastapi>=0.114",
    "uvicorn>=0.30",
    "jinja2>=3.0",
    "python-multipart>=0.0.6",
    "opencv-python>=4.12",
    "ImageHash>=4.3",
    "lpips>=0.1.4",
    "torch>=2.4",
    "torchvision>=0.19",
    "open-clip-torch>=2.26"
]
